Added snowflake anchor in configuration
Now the frozen alert icon can be on top, bottom of to the right of the information HUD. There are also two mixed modes, mixed top and mixed bottom, that draw the snowflake at the top or bottom of the block when no count is shown. Otherwise, hey will draw the snowflake next to the count text.

Other changes:
- Added translations for the snowflake anchor configuration option in `lang/en_us.json`

- Added MC command "/veinminingclient hud anchor snowflake"

- Added `SnowflakeAnchor` enum in `config.enums` & `AnchorArgumentType.SnowflakeAnchorArgumentType` type.

- Added `DrawContextMixin` & `IDrawContextMixin`

- Added an overloaded `drawText` method in DrawContext through `DrawContextMixin` & `IDrawContext` that accepts float values for X & Y instead of integers. This allows a finer control when displaying text.